Rename graph files.
