+++ src/components/plot/axis-labels.tsx
