+++ src/components/plot/plane-controls.tsx
